Refactored the existing code so that it matched the file structure we agreed upon. Also added a utils folder to the redux and created the axios with auth inside of it.
