+++ src/ReduxStuff/utils/AxiosWithAuth.js
@@ -1,0 +1,12 @@
+import axios from "axios";
+
+export const axiosWithAuth = () => {
+  const token = localStorage.getItem("token");
+
+  return axios.create({
+    baseURL: "",
+    headers: {
+      Authorization: token
+    }
+  });
+};
