+++ src/components/AccessibilityModal.js
@@ -1,7 +1,7 @@
 import React, { useState, useEffect, useRef } from "react";
 import styled from "styled-components";
 import { connect } from 'react-redux';
-import { setAdapt } from './ReduxStuff/actions';
+import { setAdapt } from '../ReduxStuff/actions/actions';
 
 const Modul = (props) => {
     console.log(props)
